mfd: intel_soc_pmic_bxtwc: Drop unneeded casting
The casting from size_t to ssize_t is not needed in addr_store(),
drop it.

Link: https://lore.kernel.org/r/20220628221747.33956-10-andriy.shevchenko@linux.intel.com

+++ drivers/mfd/intel_soc_pmic_bxtwc.c
@@ -346,7 +346,7 @@ static ssize_t addr_store(struct device *dev,
 	if (ret)
 		return ret;
 
-	return (ssize_t)count;
+	return count;
 }
 
 static ssize_t val_show(struct device *dev,
